Multi file support in eugene
Take a stab at #65 by hooking up eugene trace and eugene lint with script
discovery. It recognizes flyway style folders of versioned migrations and
sequence numbered scripts automatically. This is an early version still,
and there's lots of behavioural changes to look at (eg. should we
automatically set --commit, or refuse to trace many scripts when
it isn't set?)

The most important parameters of the current behaviour are:

- We never sort parameters provided on the command line. This means that we run
  scripts in the order they are provided.
- When provided a folder, we list all files ending in .sql and if they all have
  the same naming standard (eg. V1.*__name.sql or 13_.*.sql), we sort them in
  the correct order and run on them.
- We currently reuse the connection for eugene trace, but I think we may need
  to clean up session state to keep doing this.
- We run 1 script at the time, print the report and keep going.
- Unless there's an error, we run on all scripts before we set the exit code.