Make eugene create a temporary postgres instance
The instance is created using `initdb` and `pg_initctl` from the postgres
package on the user system, if they exist. Eugene manages the lifecycle of the
temporary server and destroys it, and its files etc when done with it.

Going to have to iterate on this a bit to figure out how it should look. This is
a nice capability, but there are lots of questions when going this route:

- It should probably be on by default, but that interacts weirdly with --commit
  which is off by default
- We have to decide if we should expose the ability to pass options to `initdb`
  and `postgres`. This complicates the UI a fair bit.