Teach eugene to filter out unmodified scripts
By passing -g ref:main, `eugene lint` and `eugene trace` can ignore scripts that
are identical on the main branch. This makes it much easier to use eugene in CI,
since it can do something useful on feature branches. `-g ref:` accepts an
arbitrary gitref expression such as a tag or `HEAD~100` etc. For now it defaults
to `-off`.